+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/omsorgspengerutbetaling-arbeidstaker-soknad
 
+## 1.9.9
+
+### Patch Changes
+
+-   Oppdatere informasjon om hvor informasjon om barn kommer fra
+-   Updated dependencies
+    -   @navikt/sif-common-ui@0.8.32
+
 ## 1.9.8
 
 ### Patch Changes