+++ apps/omsorgspengesoknad/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/omsorgspengesoknad
 
+## 1.33.10
+
+### Patch Changes
+
+-   Oppdatere informasjon om hvor informasjon om barn kommer fra
+-   Updated dependencies
+    -   @navikt/sif-common-ui@0.8.32
+
 ## 1.33.9
 
 ### Patch Changes